Implement a buffered TCP client container
